fix: promote heartbeat time_secs to i64 in thrift schema
Promote time_secs from i32 to i64 in ClusterWorkerHeartbeat,
SupervisorWorkerHeartbeat and LSWorkerHeartbeat; uptime_secs stays i32
(relative duration). Regenerate the Java and Python thrift bindings for
the three structs.

Widen the callers that the type change forces at compile time:
ExecutorBeat.timeSecs, ClusterUtils.convertExecutorBeats and the
latest-heartbeat comparison in PaceMakerStateStorage.get_worker_hb
(an absolute comparison that broke across the 2038 rollover).

Wire compat: thrift tags i32/i64 differently, so blobs written by the
old schema fail required-field validation under the new one. Heartbeats
self-heal on re-report; a full-cluster bounce upgrade is required.

+++ storm-client/src/jvm/org/apache/storm/cluster/PaceMakerStateStorage.java
@@ -149,7 +149,7 @@ public class PaceMakerStateStorage implements IStateStorage {
         while (true) {
             try {
                 byte[] ret = null;
-                int latestTimeSecs = 0;
+                long latestTimeSecs = 0;
                 boolean gotResponse = false;
 
                 HBMessage message = new HBMessage(HBServerMessageType.GET_PULSE, HBMessageData.path(path));
